requirements.txt
# Core Dependencies
python-dotenv>=1.0.1
pydantic>=2.10.6
pytz>=2024.1  # Added for timezone support
PyYAML>=6.0.1  # Added for YAML configuration files

# Web and HTTP
aiohttp>=3.9.3
requests>=2.32.3
playwright>=1.50.0
html5lib>=1.1  # Added for HTML parsing

# Search Engine
duckduckgo-search>=7.4.3

# LLM Integration
openai>=1.63.2
anthropic>=0.46.0
google-generativeai>=0.8.4
grpcio==1.60.1  # Fixed version to prevent warnings

# Testing
pytest>=8.3.4
pytest-asyncio>=0.25.3
pytest-cov>=4.1.0  # Added for coverage reports

# Data Processing and Visualization
pandas>=2.2.3
matplotlib>=3.10.0
seaborn>=0.13.2

# Utilities
tabulate>=0.9.0
tqdm>=4.67.1  # Progress bars